requirements/base.txt
django

# see https://pypi.org/project/environs/#features
environs

# django rest framwork for API and check API web view
# see https://www.django-rest-framework.org/
# https://django-rest-swagger.readthedocs.io/en/latest/
djangorestframework
django-rest-swagger

# collection of custom extensions for the Django
django-extensions

# Logging
git+https://github.com/Seedstars/django-rest-logger.git

# database
# psycopg2
# if error run 
# sudo apt-get install python3-psycopg2

# connect Firebase database 
pyrebase
